Added convenience crc32 calculated variables for Data and Data32 and used to simplify code a bit
