+++ Sources/Utils/UInt32+Crc32.swift
@@ -1,0 +1,9 @@
+//
+//  Copyright (c) 2020-2021 MobileCoin. All rights reserved.
+//
+
+import Foundation
+
+extension UInt32 {
+    static var emptyCrc32: UInt32 = 0
+}
